backend/requirements.txt
# Django 프레임워크 (v3 유지)
Django==4.2.24
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3

# 한국어 처리
hgtk>=0.2.1

# OCR + 이미지 처리
easyocr>=1.7.1
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# NLP 핵심어 추출
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
scikit-image>=0.21.0

# LLM 오케스트레이션
openai>=1.52.0
anthropic>=0.39.0
langchain>=0.3.0
langchain-anthropic>=0.2.0

# 음성 합성
gTTS>=2.5.0

# 하드웨어 통신
pyserial>=3.5

# 딥러닝 (CNN 수식 감지기)
torch>=2.0.0
torchvision>=0.15.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# 개발 도구 (선택)
ipython>=8.0.0
jupyter>=1.0.0
